Refactor code and apply single resonsibility principle Create GENERATEPEOPLE for Creating people and Listing people logic Create GENERATEBOOKS for Creating books and Listing books logic Create DATAVALUES module for holding common data between classes
