+++ data_values.rb
@@ -1,0 +1,12 @@
+module DataValues
+  @persons = []
+  @books = []
+
+  def self.persons
+    @persons
+  end
+
+  def self.books
+    @books
+  end
+end
